% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cleanSeqLevels.R
\name{cleanSeqlevels}
\alias{cleanSeqlevels}
\title{Cleans and standardizes chromosomal information.}
\usage{
cleanSeqlevels(g, excludeChr = NULL)
}
\arguments{
\item{g}{(GRanges): GRanges object to be cleaned.}

\item{excludeChr}{(character): Character vector of chromosomes which should be excluded, e.g. 'chrY' if patient is female or chrM. Set NULL to ignore.}
}
\value{
(GRanges) Returns cleaned-up GRanges.
}
\description{
Removes unused chromosomes, converts seqlevels to UCSC nomenclature and adds chromosomal information.
}
\examples{
\donttest{
 gr0 <- GenomicRanges::GRanges(
 S4Vectors::Rle(c('chr2', 'chr2', 'chr1', 'chr3'), c(1, 3, 2, 4)), 
 IRanges::IRanges(1:10, width=10:1)
 )
	gr0.clean <- cleanSeqlevels(gr0)

	GenomeInfoDb::seqinfo(gr0.clean)

}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{Misc.}
