% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_detectEnrichmentOfMutants.R
\name{detectEnrichmentOfMutants}
\alias{detectEnrichmentOfMutants}
\title{Perform a Fisher's Exact test to detect groups enriched with mutants (mutations / deep CNA).}
\usage{
detectEnrichmentOfMutants(
  mutData,
  groupInfo,
  genesOfInterest,
  minRelMutantsInGroup = 0.1
)
}
\arguments{
\item{mutData}{(tibble): Combinedreport containing the somatic mutations per gene, per sample.}

\item{groupInfo}{(tibble): Tibble containing grouping information. Needs the following columns: sample, group.}

\item{genesOfInterest}{(character): List of genes (ENSEMBL/SYMBOL) for which to always look for enrichment (regardless of minRel mutants.)}

\item{minRelMutantsInGroup}{(numeric): Only check enrichment for genes with >=X rel. mutants in the group.}
}
\value{
(tibble) Returns a tibble of the results from the Fisher's Exact test.
}
\description{
Perform a Fisher's Exact test to detect groups enriched with mutants (mutations / deep CNA).
}
