% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_detectKataegis_Old.R
\name{detectKataegis}
\alias{detectKataegis}
\title{Detect kataegis-like events by using a PCF algorithm.}
\usage{
detectKataegis(snv, min.snvs = 5, maxMeanDistance = 2000, maxk = 5)
}
\arguments{
\item{snv}{(GRanges): GRanges object containing all SNV of a single sample.}

\item{min.snvs}{(numeric): What is the minimum amount of SNV that a (constant fit) segment should have being calculated on. (How many SNV should have a consectutive fit)}

\item{maxMeanDistance}{(numeric): The maximum mean interchromosomal distance segmented SNV.}

\item{maxk}{(numeric): Max. amount of consecutive SNV which may be used to discover segmental breakpoints.}
}
\value{
(GRanges) A GRanges object containing all non-overlapping kataegis sites.
}
\description{
Uses piecewise constant fitting and additional hueristics to detect hypermutated (snvs) regions per chromosome.
}
\examples{
\dontrun{

# Example case.
detectKataegis(mutData, min.snvs = 5, maxMeanDistance = 1E3, maxk = 6)

}
}
\seealso{
Other CPCT: 
\code{\link{findCPCTCenter}()},
\code{\link{importLINXDrivers}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
