% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_determineMutationalBurden.R
\name{determineMutationalBurden}
\alias{determineMutationalBurden}
\title{Determine the mutational burden of the CPCT-02/HMF samples.}
\usage{
determineMutationalBurden(data.Cohort, minTAF.Muts = 0, minTAF.SV = 0)
}
\arguments{
\item{data.Cohort}{(list): Results of the \link[R2CPCT]{importWGSOfCohort} function.}

\item{minTAF.Muts}{(double): Min. TAF for SNV, InDel, MNV in order to count.}

\item{minTAF.SV}{(double): Min. TAF for structural variants in order to count.}
}
\value{
(tibble) Returns a tibble of the dNdS results.
}
\description{
Determine the mutational burden of the CPCT-02/HMF samples.
}
\examples{
\donttest{

 data.Cohort <- R2CPCT::importWGSOfCohort(<cpctIds>, <combinedData>)
 determineMutationalBurden(data.Cohort)

}
}
