% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_determineTiTv.R
\name{determineTiTv}
\alias{determineTiTv}
\title{Determine the number of Ti and Tv contexts and Ti/Tv ratio per sample.}
\usage{
determineTiTv(mutMatrix.SNV)
}
\arguments{
\item{mutMatrix.SNV}{(matrix): Mutational motif matrix derived from the \link[R2CPCT]{fitMutSigs} function.}
}
\value{
(tibble) Returns a tibble of the Ti/Tv number and ratio per sample.
}
\description{
Determine the number of Ti and Tv contexts and Ti/Tv ratio per sample.
}
\examples{
\donttest{

 mutSigs <- R2CPCT::fitMutSigs(data.Cohort$somaticVariants, restrictiveFit = FALSE)
 determineTiTv(mutSigs$SNV$mutMatrix)

}
}
