% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{driverList}
\alias{driverList}
\title{See misc/generateDriverList.R}
\format{
GRanges object containing the combined drivers.
}
\source{
Custom, see misc/generateDriverList.R.
}
\usage{
data(driverList)
}
\description{
See misc/generateDriverList.R
}
\keyword{datasets}
