% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_findCPCTCenter.R
\name{findCPCTCenter}
\alias{findCPCTCenter}
\title{Find corresponding center based on CPCT number.}
\usage{
findCPCTCenter(x)
}
\arguments{
\item{x}{(character): CPCT identifiers.}
}
\value{
(character) Character vector of matching centers.
}
\description{
Find corresponding center based on CPCT number.
}
\examples{
\donttest{

	meta$Center <- findCPCTCenter(CPCTId)

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis_OLD}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
