% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_fitCustomMutSigs.R
\name{fitCustomMutSigs}
\alias{fitCustomMutSigs}
\title{Performs mutational signature fitting against a pre-defined list of (custom) motif matrixes.}
\usage{
fitCustomMutSigs(dataMuts, motifMatrix)
}
\arguments{
\item{dataMuts}{(VRanges): VRanges containing the mutations which will used as input.}

\item{motifMatrix}{(tibble): Mutational motif matrix,  with a column termed 'Motif' and then x columns designating the signatures.}
}
\value{
(list) Returns a list of relevant mutational signature output.
}
\description{
Performs mutational signature fitting against a pre-defined list of (custom) motif matrixes.
}
\details{
Users need to supply a motif matrix (SNV-only), the names of the columns (signatures) will be used a aetiology.
}
\examples{
\dontrun{

 data.Cohort <- R2CPCT::importWGSOfCohort(<cpctIds>, <combinedData>)
 data.MutSigs <- R2CPCT::fitMutSigs(data.Cohort$somaticVariants, motifMatrix)

}
}
