% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisWGS_fitMutSigs.R
\name{fitMutSigs}
\alias{fitMutSigs}
\title{Performs optimal fitting of the contribution of mutational signatures.}
\usage{
fitMutSigs(dataMuts, restrictiveFit)
}
\arguments{
\item{dataMuts}{(VRanges): VRanges containing the mutations which will used as input.}

\item{restrictiveFit}{(logical): Should restrictive (strict) fitting be performed instead of using all signatures, see \link[MutationalPatterns]{fit_to_signatures_bootstrapped}?}
}
\value{
(list) Returns a list of relevant mutational signature output.
}
\description{
Performs optimal fitting of the contribution of mutational signatures.
}
\details{
Performs mutational signature fitting against the COSMIC / Alexandrov signatures (v3.1/June 2020).
It does this for SNV, InDels and MNV patterns.
}
\examples{
\dontrun{

 data.Cohort <- R2CPCT::importWGSOfCohort(<cpctIds>, <combinedData>)
 data.MutSigs <- R2CPCT::fitMutSigs(data.Cohort$somaticVariants)

}
}
