% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRNA_runArriba.R
\name{importArriba}
\alias{importArriba}
\title{Import Arriba results.}
\usage{
importArriba(inputFolder)
}
\arguments{
\item{inputFolder}{(character): Path to folder containing Arriba output.}
}
\value{
(tibble) Returns a tibble of Arriba-findings.
}
\description{
Import Arriba results.
}
\examples{
\dontrun{

	importArriba('/mnt/data2/hartwig/DR71/Oct2020/dataHMF/RNASeq/Arriba/')

}
}
\seealso{
Other RNA-Seq: 
\code{\link{retrieveDESeq2Results}()},
\code{\link{runArriba}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{RNA-Seq}
