% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_driverCatalogHMF.R
\name{importdriverCatalogHMF}
\alias{importdriverCatalogHMF}
\title{Import the driver catalog as determined by HMF.}
\usage{
importdriverCatalogHMF(pathCatalog)
}
\arguments{
\item{pathCatalog}{(character): Path to the <sample>.driver.catalog.tsv file containing HMF-determined drivers.}
}
\value{
(tibble) Tibble containing the drivers of the sample.
}
\description{
Import the driver catalog as determined by HMF.
}
\examples{
\donttest{

	importdriverCatalogHMF(pathCatalog = '<sample>.driver.catalog.tsv')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis_OLD}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
