% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_importGISTIC2.R
\name{importGISTIC2}
\alias{importGISTIC2}
\title{Imports the GISTIC2 recurrent peak and chromosomal arms output.}
\usage{
importGISTIC2(gisticFolder)
}
\arguments{
\item{gisticFolder}{(character): Path to GISTIC output folder.}
}
\value{
(list) Cleaned-up GISTIC output.
}
\description{
Imports and cleans-up the output from \link[R2CPCT]{performGISTIC2}.
}
\examples{
\dontrun{

	Folder to the GISTIC2 output.
	importGISTIC2('/tmp/etc/etc/')

}

}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
