% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_LINXDrivers.R
\name{importLINXDrivers}
\alias{importLINXDrivers}
\title{Import drivers (LOH / AMP / DEL) detected by LINX.}
\usage{
importLINXDrivers(pathDrivers)
}
\arguments{
\item{pathDrivers}{(character): Path to the <sample>.linx.drivers.tsv file containing predicted drivers.}
}
\value{
(tibble) Tibble containing the predicted drivers of the sample.
}
\description{
Import drivers (LOH / AMP / DEL) detected by LINX.
}
\examples{
\dontrun{

	importLINXDrivers(pathDrivers = '<sample>.linx.drivers.tsv')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
