% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_LINXFusions.R
\name{importLINXFusions}
\alias{importLINXFusions}
\title{Import detected fusions by LINX.}
\usage{
importLINXFusions(pathFusions)
}
\arguments{
\item{pathFusions}{(character): Path to the <sample>.linx.fusion.tsv file containing predicted fusions.}
}
\value{
(tibble) Tibble containing the predicted WGS fusions of the sample.
}
\description{
Import detected fusions by LINX.
}
\examples{
\dontrun{

	importLINXFusions(pathFusions = '<sample>.linx.fusion.tsv')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
