% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_purityStatsPURPLE.R
\name{importPurityStatsPURPLE}
\alias{importPurityStatsPURPLE}
\title{Import the driver catalog as determined by HMF.}
\usage{
importPurityStatsPURPLE(pathStats)
}
\arguments{
\item{pathStats}{(character): Path to the <sample>.purple.purity.tsv file containing HMF-determined drivers.}
}
\value{
(tibble) Tibble containing PURPLE statistics of the sample.
}
\description{
Import the driver catalog as determined by HMF.
}
\examples{
\dontrun{

	importPurityStatsPURPLE(pathStats = '<sample>.purple.purity.tsv')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
