% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_somaticCopynumbersPURPLE.R
\name{importSomaticCopynumberPURPLE}
\alias{importSomaticCopynumberPURPLE}
\title{Import the somatic CNV segments as estimated by PURPLE.}
\usage{
importSomaticCopynumberPURPLE(pathCNV)
}
\arguments{
\item{pathCNV}{(character): Path to the <sample>.purple.cnv.somatic.tsv file containing the somatic CNV regions as detected by PURPLE.}
}
\value{
(GRanges) GRanges object containing the somatic variants and assorted annotations.
}
\description{
Import the somatic CNV segments as estimated by PURPLE.
}
\details{
Will also convert the absolute copynumbers into CN-States for downstream analysis.
}
\examples{
\donttest{

	importSomaticCopynumberPURPLE(pathCNV = '<sample>.purple.cnv.somatic.tsv')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis_OLD}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
