% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_somaticVariantsVEP.R
\name{importSomaticVariantsVEP}
\alias{importSomaticVariantsVEP}
\title{Import VEP-annotated VCF files obtained from HMF (Strelka2/Sage).}
\usage{
importSomaticVariantsVEP(
  pathVCF,
  passOnly = T,
  gnomADe = 0.001,
  gnomADg = 0.005,
  keepAnnotation = T
)
}
\arguments{
\item{pathVCF}{(character): Path to the <CPCT>_post_processed.vcf file containing Strelka2 somatic variants.}

\item{passOnly}{(logical): Only import variant which passed all Strelka2 and PON filters. This also filters variants based on PON filter (>5 PON occurences).}

\item{gnomADe}{(double): Maximum allele frequency in the gnomADe database.}

\item{gnomADg}{(double): Maximum allele frequency in the gnomADg database.}

\item{keepAnnotation}{(logical): Import and interpret the ANN column?}
}
\value{
(VRanges) VRanges object containing the somatic variants and assorted annotations.
}
\description{
Imports the VCF obtained from HMF and further annotated by VEP.
}
\examples{
\dontrun{

	# Path to VEP annotated VCF.
	path <- '<CPCT#R>_<CPCT#T>_post_processed.hg19_multianno.vcf'

	sampleX.somaticVariants <- importCPCT.somaticVariants(path, passOnly = T)

	# View variant annotation.
	do.call(rbind, sampleX.somaticVariants$annotation)

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
