% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_somaticStructuralVariantsPURPLE.R
\name{importStructuralVariantsPURPLE}
\alias{importStructuralVariantsPURPLE}
\title{Import the somatic structural variants as estimated by GRIDSS / PURPLE.}
\usage{
importStructuralVariantsPURPLE(pathSV, passOnly = T)
}
\arguments{
\item{pathSV}{(character): Path to the <sample>.purple.sv.vcf.gz file containing the somatic structural variants as detected by GRIDSS / PURPLE.}

\item{passOnly}{(logical): Only retain SV which were not found in the PON and/or were inferred by copy-number transition  alone or any of the other filters.}
}
\value{
(GRanges) GRanges object containing the break-ends of the structural variants, so two rows for each SV (except when single break-end).
}
\description{
Import the somatic structural variants as estimated by GRIDSS / PURPLE.
}
\examples{
\dontrun{

	importStructuralVariantsPURPLE(pathSV = '<sample>.purple.sv.vcf.gz')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXDrivers}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
