% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_importWGSOfSample.R
\name{importWGSOfCohort}
\alias{importWGSOfCohort}
\title{Imports all the WGS data of multiple samples (cohort).}
\usage{
importWGSOfCohort(cpctIds, inputFolder, nThreads = 1, performAggregation = T)
}
\arguments{
\item{cpctIds}{(character): Full CPCT identifiers of samples, including the T suffixes (used to find files).}

\item{inputFolder}{(character): Path to the folder containing all WGS data of the samples (combinedData).}

\item{nThreads}{(integer): Number of threads to use; default is 1.}

\item{performAggregation}{(logical): Should aggregation be performed? (Is needed for further processing)}
}
\value{
(list) List containing the aggregated WGS data of the underlying samples.
}
\description{
Imports all the WGS data of multiple samples (cohort).
}
\examples{
\dontrun{

	importWGSOfCohort(cpctId = '<CPCTIDs>', 'combinedData/')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
