% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_importWGSOfSample.R
\name{importWGSOfSample}
\alias{importWGSOfSample}
\title{Imports all the WGS data of a single CPCT-02 sample.}
\usage{
importWGSOfSample(cpctId, inputFolder)
}
\arguments{
\item{cpctId}{(character): Full CPCT identifier of sample, including the T suffix (used to find files).}

\item{inputFolder}{(character): Path to the folder containing all WGS data of the sample (combinedData).}
}
\value{
(list) List containing the WGS data of the sample.
}
\description{
Imports all the WGS data of a single CPCT-02 sample.
}
\details{
Also runs CHORD for HRD predictions and ShatterSeek for chromothripsis samples.
}
\examples{
\donttest{

	importWGSOfSample(cpctId = '<CPCTID>T', 'combinedData/')

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis_OLD}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
