% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_initializeLogger.R
\name{initializeLogger}
\alias{initializeLogger}
\title{Initialize a parallel-ready logger.}
\usage{
initializeLogger(name = NULL, file = NULL)
}
\arguments{
\item{name}{(character): Name of the logger (can be empty).}

\item{file}{(character): Path to file where logging will be written to (handy when performing parallel).}
}
\value{
(NULL) Returns instance of a ParallelLogger.
}
\description{
Initialize a parallel-ready logger.
}
\examples{
\dontrun{

 initializeLogger()

}
}
