% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oncoplot.R
\name{makeOncoplot.CPCT}
\alias{makeOncoplot.CPCT}
\title{Generates a mutational signature plot.}
\usage{
makeOncoplot.CPCT(
  dataOncoplot,
  includedSamples = NULL,
  sortSamples = NULL,
  sortGenes = NULL,
  showAlleles = T
)
}
\arguments{
\item{dataOncoplot}{(tibble): Mutational data from the combinedReport.}

\item{includedSamples}{(character): Which samples should be included (completes missing samples)?}

\item{sortSamples}{(character): Specific order of samples ordering (x-axis)?}

\item{sortGenes}{(character): Specific order of genes (y-axis)?}

\item{showAlleles}{(logical): Should the mutations be (roughly) separated per paternal/maternal allele (TRUE) or combined (FALSE).}
}
\value{
(ggplot) ggplot object.
}
\description{
Generates a mutational signature plot.
}
