% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_memoSort.R
\name{memoSort}
\alias{memoSort}
\title{Sorting function based on mut. exclusivity.}
\usage{
memoSort(M)
}
\arguments{
\item{M}{(dataframe): dataframe of genes vs samples.}
}
\value{
(dataframe) Sorted dataframe of genes vs samples.
}
\description{
Sorting function based on mut. exclusivity.
}
\examples{
\donttest{

}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
