% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_performCHORD.R
\name{performCHORD}
\alias{performCHORD}
\title{Performs CHORD analysis.}
\usage{
performCHORD(somaticVariants, structuralVariants)
}
\arguments{
\item{somaticVariants}{(VRanges): VRanges of a single sample from \link[R2CPCT]{importSomaticVariantsVEP}.}

\item{structuralVariants}{(GRanges): GRanges of a single sample from \link[R2CPCT]{importStructuralVariantsPURPLE}.}
}
\value{
(list) list with the CHORD contexts and results.
}
\description{
Performs CHORD analysis.
}
\examples{
\donttest{

	runCHORD(dataSample$somaticVariants, dataSample$structuralVariants)

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis_OLD}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performShatterSeek}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
