% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_performGISTIC2.R
\name{performGISTIC2}
\alias{performGISTIC2}
\title{Detect genes under selection by focal amplification/deletion using the GISTIC2.0 algorithm.}
\usage{
performGISTIC2(
  regions,
  outputFolder,
  gisticPath = "/mnt/onco0002/repository/software/GISTIC2/gistic2",
 
    gisticParameters = "-genegistic 1 -gcm extreme -maxseg 4000 -broad 1 -brlen 0.98 -conf 0.95 -rx 0 -cap 3 -saveseg 0 -armpeel 1 -smallmem 0 -res 0.01 -ta 0.3 -td 0.3 -savedata 0 -savegene 1 -qvt 0.1 -twoside 0",
 
    gisticAnno = "/mnt/onco0002/repository/software/GISTIC2/refgenefiles/hg19.UCSC.add_miR.140312.refgene.mat"
)
}
\arguments{
\item{regions}{(GRanges): GRanges of CN segments with the following columns: sample, copyNumber, bafCount, sex.}

\item{outputFolder}{(character): Path to output folder of GISTIC2 results.}

\item{gisticPath}{(character): Path to gistic2 shell script.}

\item{gisticParameters}{(character): Optional parameters given to gistic2. See online documentation for options.}

\item{gisticAnno}{(character): Path to which genome annotation GISTIC should use? Stored in /mnt/data/ccbc_environment/software/general/GISTIC2_2.0.23/}
}
\value{
(list) GISTIC
}
\description{
This can take anywhere between 30 min. up to several hours depending on the number of samples.
If you have previously ran GISTIC, use the \link[R2CPCT]{importGISTIC2} function to import and clean the data.
}
\examples{
\dontrun{

	gisticOutput <- R2CPCT::performGISTIC2(regions = data.Cohort$copyNumbers)

}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{GISTIC}
