% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_performLINX.R
\name{performLINX}
\alias{performLINX}
\title{Run LINX on one or multiple samples to discover fusion and driver-events from PURPLE output.}
\usage{
performLINX(pathCombined, cpctIds, nThreads, dryRun = T)
}
\arguments{
\item{pathCombined}{(character): Path to the combinedData/ folder containing the symlinks to the WGS files of the data-request (DR).}

\item{cpctIds}{(character): The samples which will be retrieved.}

\item{nThreads}{(integer): Number of cores to run multiple LINX in parallel.}

\item{dryRun}{(logical): Only output the tmp. file which stores the commands (TRUE) or run the commands (FALSE).}
}
\value{
(NULL) Does not return anything, only outputs logging messages.
}
\description{
For installation instructions, please see the 'howToInstallLINX' vignette.
It will output the (relevant) files in the combinedData/ folder.
}
\examples{
\dontrun{

 runLINX('DR-071/combinedData/', c('CPCT02124567', 'CPCT02012345123'), nThreads = 10)

}
}
