% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_performShatterSeek.R
\name{performShatterSeek}
\alias{performShatterSeek}
\title{Performs ShatterSeek analysis.}
\usage{
performShatterSeek(copyNumbers, structuralVariants)
}
\arguments{
\item{copyNumbers}{(GRanges): VRanges of a single sample from \link[R2CPCT]{importSomaticVariantsVEP}.}

\item{structuralVariants}{(GRanges): GRanges of a single sample from \link[R2CPCT]{importStructuralVariantsPURPLE}.}
}
\value{
(list) Tibble with the CHORD results.
}
\description{
Performs ShatterSeek analysis.
}
\examples{
\dontrun{

	runCHORD(dataSample$somaticVariants, dataSample$structuralVariants)

}
}
\seealso{
Other CPCT: 
\code{\link{detectKataegis}()},
\code{\link{findCPCTCenter}()},
\code{\link{importLINXFusions}()},
\code{\link{importPurityStatsPURPLE}()},
\code{\link{importSomaticCopynumberPURPLE}()},
\code{\link{importSomaticVariantsVEP}()},
\code{\link{importStructuralVariantsPURPLE}()},
\code{\link{importWGSOfCohort}()},
\code{\link{importWGSOfSample}()},
\code{\link{importdriverCatalogHMF}()},
\code{\link{performCHORD}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{CPCT}
