% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_performVEP.R
\name{performVEP}
\alias{performVEP}
\title{Generate the commands to run VEP on HMF/CPCT-02 samples.}
\usage{
performVEP(
  pathVCF,
  pathOutput = NULL,
  cpctIds,
 
    fastaFile = "/mnt/data/ccbc_environment/general/genomes/hsapiens/hg19_HMF/Homo_sapiens.GRCh37.GATK.illumina.fasta",
 
    pathGTF = "/mnt/data/ccbc_environment/general/annotation/hg19/GENCODE/noChrPrefix_VEP_gencode.v35lift37.annotation.gtf.gz"
)
}
\arguments{
\item{pathVCF}{(character): Path to folder containing the CPCT-02 VCF files (combinedData/<sample>.purple.somatic.vcf.gz).}

\item{pathOutput}{(character): Path to output folder for VEP-annotated VCF files (if left empty, pathVCF will be used).}

\item{cpctIds}{(character): The samples which will be annotated by VEP.}

\item{fastaFile}{(character): Path to genome FASTA.}

\item{pathGTF}{(character): Path to GTF (see vignette on VEP install).}
}
\value{
(character) Returns the command which should be performed in bash.
}
\description{
Generate the bash commands to run VEP on HMF/CPCT-02 samples.
Please see the vignette 'howToInstallVep' on instructions to install VEP.

The accompanying bash script is provided in inst/performVEPInParallel.sh.
This function will return the command which should be run in bash.

By default, it performs annotation based on ensembl-vep (v101) and GENCODE v35 (hg19) gene-structures.
}
\examples{
\dontrun{

 performVEP('DR-071/combinedData/', cpctIds = c('CPCT02124567', 'CPCT02012345123'))

}
}
