% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MutationalSignaturesCOSMIC.R
\name{plotMutationalSignaturesCOSMIC}
\alias{plotMutationalSignaturesCOSMIC}
\title{Generates a mutational signature plot.}
\usage{
plotMutationalSignaturesCOSMIC(
  mutSigs,
  orderSamples = NULL,
  minContrib = 5,
  minMutants = NULL,
  combineSigs = T,
  focusOnSig = NULL,
  onlyShowFocus = F
)
}
\arguments{
\item{mutSigs}{(list): A list containing the data of the mutational reconstruction (of either the SNV, InDel or DBS) from \code{\link{fitMutSigs}}.}

\item{orderSamples}{(character): Specific ordering of samples.}

\item{minContrib}{(numeric): Min. contribution of a signature (in any sample) in order to plot it. Set to 0 to plot all possible signatures.}

\item{minMutants}{(numeric): Min. number of mutants (SNV, DBS or InDel) on which the fitting should have been performed.
Else, it will be flagged as a low-confident reconstruction.}

\item{combineSigs}{(logical): Combine signatures based on aetiology?}

\item{focusOnSig}{(character): Should there be a focus on specific signatures? All others will be categorized as 'Other'.}

\item{onlyShowFocus}{(logical): Should only the focused signatures be shown?}
}
\value{
(ggplot) ggplot object.
}
\description{
Generates a mutational signature plot.
}
