% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overviewGISTIC2.R
\name{plotOverviewGISTIC2}
\alias{plotOverviewGISTIC2}
\title{Generates an overview of the GISTIC2 copy-number landscape.}
\usage{
plotOverviewGISTIC2(
  data.GISTIC,
  maxG = 3,
  minG = -2,
  maxQ = 0.1,
  sizeTrackCN = 15
)
}
\arguments{
\item{data.GISTIC}{(list): Output from \code{\link{importGISTIC2}}.}

\item{maxG}{(numeric): Cut-off of the max. G-Score.}

\item{minG}{(numeric): Cut-off of the min. G-Score.}

\item{maxQ}{(numeric): Which peaks are plotted (max. q-value).}

\item{sizeTrackCN}{(numeric): Size of the CN track.}
}
\value{
(NULL) Plots a Circos figure.
}
\description{
Generates an overview of the GISTIC2 copy-number landscape.
}
\examples{
\donttest{

 svg(file = 'overviewCNA.svg', width = 8, height = 8, onefile = TRUE)
 plotGISTIC2(data.GISTIC, 2, ,-2, 0.1)
 dev.off()

}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
