% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposedAetiologyCOSMICv3.2}
\alias{proposedAetiologyCOSMICv3.2}
\title{library(rvest)}
\format{
Tibble containing the COSMIC (v3.2; March 2021) proposed signatures aetiologies.
}
\source{
COSMIC
}
\usage{
data(proposedAetiologyCOSMICv3.2)
}
\description{
# SBS Signatures ----------------------------------------------------------
}
\details{
URL <- 'https://cancer.sanger.ac.uk/cosmic/signatures/SBS/'

# Retrieve content.
x <- xml2::read_html(URL) %>%
  html_nodes(css = '.vignettes_table_plot') %>%
  html_children() %>%
  html_text(trim = TRUE)

sigs.SBS <- rbind(
  do.call(rbind, lapply(x, function(s){
    s <- trimws(strsplit(s, '\n')[[1]])
    data.frame(Signature = s[[1]], proposedAetiology = gsub('Proposed Aetiology', '', s[[6]]))
  }))
  , data.frame(Signature = c('SBS27', 'SBS43', 'SBS45','SBS46','SBS47','SBS48','SBS49','SBS50','SBS51','SBS52','SBS53','SBS54','SBS55','SBS56','SBS57','SBS58','SBS59','SBS60'), proposedAetiology = 'Possible Sequencing Artifact')
)


# DBS Signatures ----------------------------------------------------------

URL <- 'https://cancer.sanger.ac.uk/cosmic/signatures/DBS/'

# Retrieve content.
x <- xml2::read_html(URL) %>%
  html_nodes(css = '.vignettes_table_plot') %>%
  html_children() %>%
  html_text(trim = TRUE)

sigs.DBS <- do.call(rbind, lapply(x, function(s){
  s <- trimws(strsplit(s, '\n')[[1]])
  data.frame(Signature = s[[1]], proposedAetiology = gsub('Proposed Aetiology', '', s[[6]]))
}))

# InDel Signatures --------------------------------------------------------

URL <- 'https://cancer.sanger.ac.uk/cosmic/signatures/ID'

# Retrieve content.
x <- xml2::read_html(URL) %>%
  html_nodes(css = '.vignettes_table_plot') %>%
  html_children() %>%
  html_text(trim = TRUE)

sigs.ID <- do.call(rbind, lapply(x, function(s){
  s <- trimws(strsplit(s, '\n')[[1]])
  data.frame(Signature = s[[1]], proposedAetiology = gsub('Proposed Aetiology', '', s[[6]]))
}))


# Combine lists -----------------------------------------------------------

proposedAetiologyCOSMICv3.2 <- rbind(sigs.SBS, sigs.DBS, sigs.ID)

# Group signatures based on common proposed aetiology together.
proposedAetiologyCOSMICv3.2 <- proposedAetiologyCOSMICv3.2 %>%
  dplyr::group_by(proposedAetiology) %>%
  dplyr::mutate(proposedAetiologyGrouped = sprintf('%s (%s)', proposedAetiology, ifelse(dplyr::n_distinct(Signature) <= 4, paste(Signature, collapse = ', '), '>4 signatures'))) %>%
  dplyr::ungroup()

usethis::use_data(proposedAetiologyCOSMICv3.2)
}
\keyword{datasets}
