% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteinDomains}
\alias{proteinDomains}
\title{# Import protein domains from PROT2HG --------------------------------------}
\format{
GRanges containing protein domains and protein-specific sites, derived from PROT2HG (Jan. 2021)
}
\source{
proteins
}
\usage{
data(proteinDomains)
}
\description{
proteinDomains <- data.table::fread(input = '~/prot2hg_1938_112019.csv', sep = ';', stringsAsFactors = FALSE)
}
\details{
# Select protein-information.
proteinDomains <- proteinDomains %>% dplyr::select(protein_ID, gene, prot_start, prot_end, type, feature_name, note, ensembl) %>% dplyr::distinct()

# Convert to GRanges.
proteinDomains <- GenomicRanges::makeGRangesFromDataFrame(proteinDomains, keep.extra.columns = T, seqnames.field = 'protein_ID', start.field = 'prot_start', end.field = 'prot_end')

# Factorize to reduce size.
proteinDomains$type <- factor(proteinDomains$type)
proteinDomains$gene <- factor(proteinDomains$gene)
proteinDomains$ensembl <- factor(proteinDomains$ensembl)

usethis::use_data(proteinDomains)
}
\keyword{datasets}
