% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRNA_retrieveDESeq2Results.R
\name{retrieveDESeq2Results}
\alias{retrieveDESeq2Results}
\title{Retrieve the results from a DESeq2-analysis using IHW and LFC shrinkage.}
\usage{
retrieveDESeq2Results(DESeq2.Object, contrast, nThreads = 20)
}
\arguments{
\item{DESeq2.Object}{(DESeqDataSet): DESeqDataSet of a DESeq2 analysis.}

\item{contrast}{(character): Contrast of the resulst, e.g.: c('testVariable', 'groupA', 'groupb')}

\item{nThreads}{(integer): Number of threads using during multi-threading}
}
\value{
(tibble) Tibble containing the DESeq2 results.
}
\description{
Retrieve the results from a DESeq2-analysis using IHW and LFC shrinkage.
}
\examples{
\dontrun{

DESeq2.allSamples <- DESeq2::DESeqDataSetFromMatrix(countData = countMatrix, colData = samples.meta[match(colnames(countMatrix), samples.meta$sample),], design = ~treatmentType)
DESeq2.allSamples <- DESeq2::DESeq(DESeq2.allSamples, test = 'Wald', parallel = T, BPPARAM = BiocParallel::MulticoreParam(workers = 20))
retrieveDESeq2Results(DESeq2.allSamples)

}
}
\seealso{
Other RNA-Seq: 
\code{\link{importArriba}()},
\code{\link{runArriba}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{RNA-Seq}
