% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_retrieveFilesWGS.R
\name{retrieveFilesWGS}
\alias{retrieveFilesWGS}
\title{Retrieves the WGS files needed for analysis into the same folder.}
\usage{
retrieveFilesWGS(pathHMF, pathOutput, cpctIds)
}
\arguments{
\item{pathHMF}{(character): Path to main folder containing the WGS files of the data-request (DR).}

\item{pathOutput}{(character): Path to output folder which will contain the symlinks.}

\item{cpctIds}{(character): The samples which will be retrieved.}
}
\value{
(NULL) Does not return anything, only outputs logging messages.
}
\description{
Retrieves the files (grep) based on CPCT identifiers and symlinks them into the same folder for easier manipulation.
}
\examples{
\donttest{

 retrieveFilesWGS(
 pathHMF = 'DR-071/rawData/WGS/',
 pathOutput = 'DR-071/combinedData/',
 cpctIds = c('CPCT02124567', 'CPCT02012345123')
 )

}
}
