% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRNA_runArriba.R
\name{runArriba}
\alias{runArriba}
\title{Run Arriba on RNA-Seq samples using structural-variants information derived from WGS.}
\usage{
runArriba(data.SV, minTAF = 0.05, bamFolder, outputFolder)
}
\arguments{
\item{data.SV}{(GRanges): GRanges of SV derived from WGS.}

\item{minTAF}{(integer): Min. TAF of SV to take into the analysis, set to zero (0) for no filtering.}

\item{bamFolder}{(character): Path to folder containing the RNA-Seq BAM files (STAR).}

\item{outputFolder}{(character): Path to folder to write Arriba output.}
}
\value{
(chr) Command to run Arriba.
}
\description{
Run Arriba on RNA-Seq samples using structural-variants information derived from WGS.
}
\examples{
\donttest{

	runArriba(
	  data.SV = DR71.CohortWGS$structuralVariants,
	  minTAF = 0.05,
	  bamFolder = '/mnt/data2/hartwig/DR71/Oct2020/dataHMF/RNASeq/BAM/',
	  outputFolder = '/mnt/data2/hartwig/DR71/Oct2020/dataHMF/RNASeq/Arriba/'
	)

}
}
\seealso{
Other RNA-Seq: 
\code{\link{importArriba}()},
\code{\link{retrieveDESeq2Results}()}
}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}
}
\concept{RNA-Seq}
